Uploaded changes of websocket of terminal.
We removed the config of websocket to the new router.
